requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

python-dotenv
langchain
langchain_core
langgraph
langchainhub
langchain-community
langchain_openai
langchain-tavily
langchain_postgres
langgraph-checkpoint-postgres

beautifulsoup4
tiktoken

pytest
pytest-asyncio

httpx
